Trying to get slices to work.
